// src/api/graphql/types.generated.ts
export type Maybe<T> = T | null | undefined;
// ...
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
	DateTime: { input: Date; output: Date };
};

export type Cart = {
	__typename?: "Cart";
	id: Scalars["String"]["output"];
	items?: Maybe<Array<Maybe<CartItem>>>;
	order?: Maybe<Order>;
};

export type CartInput = {
	items?: InputMaybe<Array<InputMaybe<CartItemInput>>>;
};

export type CartItem = {
	__typename?: "CartItem";
	cart?: Maybe<Cart>;
	id: Scalars["String"]["output"];
	product: Product;
	quantity: Scalars["Int"]["output"];
};

export type CartItemInput = {
	cartId: Scalars["String"]["input"];
	productId: Scalars["String"]["input"];
	quantity: Scalars["Int"]["input"];
};

export type Category = {
	__typename?: "Category";
	description: Scalars["String"]["output"];
	id: Scalars["ID"]["output"];
	name: Scalars["String"]["output"];
	products?: Maybe<Array<Maybe<Product>>>;
	slug: Scalars["String"]["output"];
};

export type CategoryInput = {
	description: Scalars["String"]["input"];
	name: Scalars["String"]["input"];
};

export type Collection = {
	__typename?: "Collection";
	description: Scalars["String"]["output"];
	id: Scalars["ID"]["output"];
	name: Scalars["String"]["output"];
	products?: Maybe<Array<Maybe<Product>>>;
	slug: Scalars["String"]["output"];
};

export type CollectionInput = {
	description: Scalars["String"]["input"];
	name: Scalars["String"]["input"];
};

export type Image = {
	__typename?: "Image";
	alt: Scalars["String"]["output"];
	height: Scalars["Int"]["output"];
	id: Scalars["String"]["output"];
	url: Scalars["String"]["output"];
	width: Scalars["Int"]["output"];
};

export type ImageInput = {
	alt: Scalars["String"]["input"];
	height: Scalars["Int"]["input"];
	productId: Scalars["String"]["input"];
	url: Scalars["String"]["input"];
	width: Scalars["Int"]["input"];
};

export type Mutation = {
	__typename?: "Mutation";
	createCart: Cart;
	createCartItem: CartItem;
	createCategory: Category;
	createCollection: Collection;
	createImage: Image;
	createOrder: Order;
	createProduct: Product;
	createReview: Review;
	deleteCart: Cart;
	deleteCartItem: Cart;
	deleteCategory: Category;
	deleteCollection: Collection;
	deleteImage: Image;
	deleteOrder: Order;
	deleteProduct: Product;
	deleteReview: Review;
	updateCart: Cart;
	updateCartItem: CartItem;
	updateCategory: Category;
	updateCollection: Collection;
	updateImage: Image;
	updateOrder: Order;
	updateProduct: Product;
	updateReview: Review;
};
// ...
export type Order = {
	__typename?: "Order";
	createdAt: Scalars["String"]["output"];
	email: Scalars["String"]["output"];
	id: Scalars["String"]["output"];
	items?: Maybe<Scalars["String"]["output"]>;
	status: Scalars["String"]["output"];
	total: Scalars["Int"]["output"];
	updatedAt: Scalars["String"]["output"];
};

export type OrderInput = {
	cartId: Scalars["String"]["input"];
	email: Scalars["String"]["input"];
	status: Scalars["String"]["input"];
	total: Scalars["Int"]["input"];
};

export type OrderStatus =
	| "CANCELLED"
	| "COMPLETED"
	| "CREATED"
	| "PAID"
	| "PROCESSING";

export type Product = {
	__typename?: "Product";
	categories?: Maybe<Array<Maybe<Category>>>;
	collections?: Maybe<Array<Maybe<Collection>>>;
	description: Scalars["String"]["output"];
	id: Scalars["ID"]["output"];
	images?: Maybe<Array<Maybe<Image>>>;
	name: Scalars["String"]["output"];
	price: Scalars["Int"]["output"];
	reviews?: Maybe<Array<Maybe<Review>>>;
};

export type ProductInput = {
	categoryIds?: InputMaybe<Array<InputMaybe<Scalars["ID"]["input"]>>>;
	collectionIds?: InputMaybe<Array<InputMaybe<Scalars["ID"]["input"]>>>;
	description: Scalars["String"]["input"];
	imageIds?: InputMaybe<Array<InputMaybe<Scalars["ID"]["input"]>>>;
	name: Scalars["String"]["input"];
	price: Scalars["Int"]["input"];
};

export type Query = {
	__typename?: "Query";
	cart?: Maybe<Cart>;
	cartitem?: Maybe<CartItem>;
	cartitems: Array<CartItem>;
	carts: Array<Cart>;
	categories?: Maybe<Array<Maybe<Category>>>;
	category?: Maybe<Category>;
	collection?: Maybe<Collection>;
	collections?: Maybe<Array<Maybe<Collection>>>;
	image?: Maybe<Image>;
	images: Array<Image>;
	order?: Maybe<Order>;
	orders: Array<Order>;
	product?: Maybe<Product>;
	products?: Maybe<Array<Maybe<Product>>>;
	review?: Maybe<Review>;
	reviews: Array<Review>;
};
// ...
export type Review = {
	__typename?: "Review";
	author: Scalars["String"]["output"];
	content: Scalars["String"]["output"];
	createdAt: Scalars["DateTime"]["output"];
	id: Scalars["ID"]["output"];
	rating: Scalars["Int"]["output"];
	title: Scalars["String"]["output"];
};

export type ReviewInput = {
	author: Scalars["String"]["input"];
	content: Scalars["String"]["input"];
	productId: Scalars["ID"]["input"];
	rating: Scalars["Int"]["input"];
	title: Scalars["String"]["input"];
};

export type ResolverTypeWrapper<T> = Promise<T> | T;

export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
	resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
	| ResolverFn<TResult, TParent, TContext, TArgs>
	| ResolverWithResolve<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
	parent: TParent,
	args: TArgs,
	context: TContext,
	info?: GraphQLResolveInfo,
) => Promise<TResult> | TResult;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (
	obj: T,
	context: TContext,
	info?: GraphQLResolveInfo,
) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<
	TResult = {},
	TParent = {},
	TContext = {},
	TArgs = {},
> = (
	next: NextResolverFn<TResult>,
	parent: TParent,
	args: TArgs,
	context: TContext,
	info?: GraphQLResolveInfo,
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
	Cart: ResolverTypeWrapper<Mapper<Cart>>;
	String: ResolverTypeWrapper<Mapper<Scalars["String"]["output"]>>;
	CartInput: ResolverTypeWrapper<Mapper<CartInput>>;
	CartItem: ResolverTypeWrapper<Mapper<CartItem>>;
	Int: ResolverTypeWrapper<Mapper<Scalars["Int"]["output"]>>;
	CartItemInput: ResolverTypeWrapper<Mapper<CartItemInput>>;
	Category: ResolverTypeWrapper<Mapper<Category>>;
	ID: ResolverTypeWrapper<Mapper<Scalars["ID"]["output"]>>;
	CategoryInput: ResolverTypeWrapper<Mapper<CategoryInput>>;
	Collection: ResolverTypeWrapper<Mapper<Collection>>;
	CollectionInput: ResolverTypeWrapper<Mapper<CollectionInput>>;
	DateTime: ResolverTypeWrapper<Mapper<Scalars["DateTime"]["output"]>>;
	Image: ResolverTypeWrapper<Mapper<Image>>;
	ImageInput: ResolverTypeWrapper<Mapper<ImageInput>>;
	Mutation: ResolverTypeWrapper<{}>;
	Order: ResolverTypeWrapper<Mapper<Order>>;
	OrderInput: ResolverTypeWrapper<Mapper<OrderInput>>;
	OrderStatus: ResolverTypeWrapper<Mapper<OrderStatus>>;
	Product: ResolverTypeWrapper<Mapper<Product>>;
	ProductInput: ResolverTypeWrapper<Mapper<ProductInput>>;
	Query: ResolverTypeWrapper<{}>;
	Review: ResolverTypeWrapper<Mapper<Review>>;
	ReviewInput: ResolverTypeWrapper<Mapper<ReviewInput>>;
	Boolean: ResolverTypeWrapper<Mapper<Scalars["Boolean"]["output"]>>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
	Cart: Mapper<Cart>;
	String: Mapper<Scalars["String"]["output"]>;
	CartInput: Mapper<CartInput>;
	CartItem: Mapper<CartItem>;
	Int: Mapper<Scalars["Int"]["output"]>;
	CartItemInput: Mapper<CartItemInput>;
	Category: Mapper<Category>;
	ID: Mapper<Scalars["ID"]["output"]>;
	CategoryInput: Mapper<CategoryInput>;
	Collection: Mapper<Collection>;
	CollectionInput: Mapper<CollectionInput>;
	DateTime: Mapper<Scalars["DateTime"]["output"]>;
	Image: Mapper<Image>;
	ImageInput: Mapper<ImageInput>;
	Mutation: {};
	Order: Mapper<Order>;
	OrderInput: Mapper<OrderInput>;
	Product: Mapper<Product>;
	ProductInput: Mapper<ProductInput>;
	Query: {};
	Review: Mapper<Review>;
	ReviewInput: Mapper<ReviewInput>;
	Boolean: Mapper<Scalars["Boolean"]["output"]>;
};
// ...
export type ProductResolvers<
	ContextType = Context,
	ParentType extends ResolversParentTypes["Product"] = ResolversParentTypes["Product"],
> = {
	categories?: Resolver<
		Maybe<Array<Maybe<ResolversTypes["Category"]>>>,
		ParentType,
		ContextType
	>;
	collections?: Resolver<
		Maybe<Array<Maybe<ResolversTypes["Collection"]>>>,
		ParentType,
		ContextType
	>;
	description?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
	id?: Resolver<ResolversTypes["ID"], ParentType, ContextType>;
	images?: Resolver<
		Maybe<Array<Maybe<ResolversTypes["Image"]>>>,
		ParentType,
		ContextType
	>;
	name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
	price?: Resolver<ResolversTypes["Int"], ParentType, ContextType>;
	reviews?: Resolver<
		Maybe<Array<Maybe<ResolversTypes["Review"]>>>,
		ParentType,
		ContextType
	>;
	__isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};